+++ assets/javascripts/kitten/components/buttons/button-image-with-text/test.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { ButtonImageWithText } from '../../components/buttons/button-image-with-text'
+import { ButtonImageWithText } from './index'
 
 describe('<ButtonImageWithText />', () => {
-  it('should match its empty snapshot', () => {
+  it('should match with snapshot', () => {
     const tree = renderer
       .create(
         <ButtonImageWithText
@@ -13,6 +13,9 @@ describe('<ButtonImageWithText />', () => {
           largeGutter={false}
           titleRegular={false}
           textRegular={false}
+          avatarProps={{
+            src: 'https://placehold.it/100x100/caf4fe/caf4fe',
+          }}
         />,
       )
       .toJSON()